@@ -7,67 +7,63 @@
   "repository": "transitive-bullshit/nextjs-notion-starter-kit",
   "license": "MIT",
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "deploy": "vercel --prod",
+    "deploy": "vercel deploy",
     "deps": "run-s deps:*",
     "deps:update": "[ -z $GITHUB_ACTIONS ] && yarn add notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
     "deps:link": "[ -z $GITHUB_ACTIONS ] && yarn link notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:link on CI'",
     "analyze": "cross-env ANALYZE=true next build",
     "analyze:server": "cross-env BUNDLE_ANALYZE=server next build",
     "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build",
-    "test": "run-s test:*",
-    "test:lint": "eslint .",
-    "test:prettier": "prettier '**/*.{js,jsx,ts,tsx}' --check",
-    "posttest": "run-s build"
+    "test": "run-p test:*",
+    "test:lint": "eslint '**/*.{ts,tsx}'",
+    "test:prettier": "prettier '**/*.{js,jsx,ts,tsx}' --check"
   },
   "dependencies": {
-    "@google-cloud/firestore": "^4.10.1",
+    "@keyv/redis": "^2.2.3",
     "classnames": "^2.3.1",
     "date-fns": "^2.25.0",
+    "expiry-map": "^2.0.0",
     "fathom-client": "^3.0.0",
     "got": "^11.8.2",
     "isomorphic-unfetch": "^3.1.0",
+    "keyv": "^4.1.1",
     "lqip-modern": "^1.2.0",
-    "next": "^11.1.2",
+    "next": "^12.1.0",
     "node-fetch": "^2.6.1",
-    "notion-client": "^4.10.0",
-    "notion-types": "^4.10.0",
-    "notion-utils": "^4.10.0",
-    "p-map": "^4.0.0",
-    "p-memoize": "^4.0.0",
+    "notion-client": "^6.0.8",
+    "notion-types": "^6.0.6",
+    "notion-utils": "^6.0.8",
+    "p-map": "^5.3.0",
+    "p-memoize": "^6.0.1",
     "react": "^17.0.2",
     "react-body-classname": "^1.3.1",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
-    "react-notion-x": "^4.11.0",
-    "react-static-tweets": "^0.5.3",
-    "react-use": "^17.2.4",
-    "static-tweets": "^0.5.3",
+    "react-notion-x": "^6.0.8",
+    "react-static-tweets": "^0.7.1",
+    "react-use": "^17.3.2",
+    "static-tweets": "^0.7.1",
     "use-dark-mode": "^2.3.1"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^11.1.2",
+    "@next/bundle-analyzer": "^12.1.0",
     "@types/classnames": "^2.3.1",
     "@types/node": "^16.11.2",
     "@types/node-fetch": "^3.0.3",
     "@types/react": "^17.0.31",
-    "@typescript-eslint/eslint-plugin": "^5.1.0",
-    "@typescript-eslint/parser": "^5.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.15.0",
+    "@typescript-eslint/parser": "^5.15.0",
     "cross-env": "^7.0.2",
-    "eslint": "^8.0.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-config-standard-react": "^11.0.1",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-promise": "^5.1.1",
-    "eslint-plugin-react": "^7.26.1",
+    "eslint": "^8.11.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
     "typescript": "^4.4.4"
