@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   // where it all starts -- the site's root Notion page (required)
   rootNotionPageId: '78fc5a4b88d74b0e824e29407e9f1ec1',
 
@@ -27,17 +27,13 @@ module.exports = {
   defaultPageCover: null,
   defaultPageCoverPosition: 0.5,
 
-  // image CDN host to proxy all image requests through (optional)
-  // NOTE: this requires you to set up an external image proxy
-  imageCDNHost: null,
-
-  // Utteranc.es comments via GitHub issue comments (optional)
-  utterancesGitHubRepo: null,
-
   // whether or not to enable support for LQIP preview images (optional)
-  // NOTE: this requires you to set up Google Firebase and add the environment
-  // variables specified in .env.example
-  isPreviewImageSupportEnabled: false,
+  isPreviewImageSupportEnabled: true,
+
+  // whether or not redis is enabled for caching generated preview images (optional)
+  // NOTE: if you enable redis, you need to set the `REDIS_HOST` and `REDIS_PASSWORD`
+  // environment variables. see the readme for more info
+  isRedisEnabled: false,
 
   // map of notion page IDs to URL paths (optional)
   // any pages defined here will override their default URL paths
