@@ -1,9 +1,12 @@
 import { NotionAPI } from 'notion-client'
 import { ExtendedRecordMap, SearchParams, SearchResults } from 'notion-types'
-import { getPreviewImages } from './get-preview-images'
-import { mapNotionImageUrl } from './map-image-url'
-import { fetchTweetAst } from 'static-tweets'
-import pMap from 'p-map'
+
+import { getPreviewImageMap } from './preview-images'
+import { getTweetAstMap } from './tweet-embeds'
+import {
+  isPreviewImageSupportEnabled,
+  isTweetEmbedSupportEnabled
+} from './config'
 
 export const notion = new NotionAPI({
   apiBaseUrl: process.env.NOTION_API_BASE_URL
@@ -11,92 +14,16 @@ export const notion = new NotionAPI({
 
 export async function getPage(pageId: string): Promise<ExtendedRecordMap> {
   const recordMap = await notion.getPage(pageId)
-  const blockIds = Object.keys(recordMap.block)
 
-  const imageUrls: string[] = blockIds
-    .map((blockId) => {
-      const block = recordMap.block[blockId]?.value
+  if (isPreviewImageSupportEnabled) {
+    const previewImageMap = await getPreviewImageMap(recordMap)
+    ;(recordMap as any).preview_images = previewImageMap
+  }
 
-      if (block) {
-        if (block.type === 'image') {
-          const source = block.properties?.source?.[0]?.[0]
-
-          if (source) {
-            return {
-              block,
-              url: source
-            }
-          }
-        }
-
-        if ((block.format as any)?.page_cover) {
-          const source = (block.format as any).page_cover
-
-          return {
-            block,
-            url: source
-          }
-        }
-      }
-
-      return null
-    })
-    .filter(Boolean)
-    .map(({ block, url }) => mapNotionImageUrl(url, block))
-    .filter(Boolean)
-
-  const urls = Array.from(new Set(imageUrls))
-  const previewImageMap = await getPreviewImages(urls)
-  ;(recordMap as any).preview_images = previewImageMap
-
-  const tweetIds: string[] = blockIds
-    .map((blockId) => {
-      const block = recordMap.block[blockId]?.value
-
-      if (block) {
-        if (block.type === 'tweet') {
-          const src = block.properties?.source?.[0]?.[0]
-
-          if (src) {
-            const id = src.split('?')[0].split('/').pop()
-            if (id) return id
-          }
-        }
-      }
-
-      return null
-    })
-    .filter(Boolean)
-
-  const tweetAsts = await pMap(
-    tweetIds,
-    async (tweetId) => {
-      try {
-        return {
-          tweetId,
-          tweetAst: await fetchTweetAst(tweetId)
-        }
-      } catch (err) {
-        console.error('error fetching tweet info', tweetId, err)
-      }
-    },
-    {
-      concurrency: 4
-    }
-  )
-
-  const tweetAstMap = tweetAsts.reduce((acc, { tweetId, tweetAst }) => {
-    if (tweetAst) {
-      return {
-        ...acc,
-        [tweetId]: tweetAst
-      }
-    } else {
-      return acc
-    }
-  }, {})
-
-  ;(recordMap as any).tweetAstMap = tweetAstMap
+  if (isTweetEmbedSupportEnabled) {
+    const tweetAstMap = await getTweetAstMap(recordMap)
+    ;(recordMap as any).tweetAstMap = tweetAstMap
+  }
 
   return recordMap
 }
