@@ -1,11 +1,19 @@
-// const isDev = process.env.NODE_ENV === 'development' || !process.env.NODE_ENV
-
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const withBundleAnalyzer = require('@next/bundle-analyzer')({
   enabled: process.env.ANALYZE === 'true'
 })
 
 module.exports = withBundleAnalyzer({
+  staticPageGenerationTimeout: 300,
   images: {
-    domains: ['pbs.twimg.com']
+    domains: [
+      'www.notion.so',
+      'notion.so',
+      'images.unsplash.com',
+      'pbs.twimg.com',
+      'abs.twimg.com',
+      'transitivebullsh.it'
+    ],
+    formats: ['image/avif', 'image/webp']
   }
 })
