@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next'
 
-import * as types from '../lib/types'
-import { search } from '../lib/notion'
+import * as types from '../../lib/types'
+import { search } from '../../lib/notion'
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
   if (req.method !== 'POST') {
@@ -10,8 +10,9 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
 
   const searchParams: types.SearchParams = req.body
 
-  console.log('lambda search-notion', searchParams)
+  console.log('<<< lambda search-notion', searchParams)
   const results = await search(searchParams)
+  console.log('>>> lambda search-notion', results)
 
   res.setHeader(
     'Cache-Control',
